Fixed a Windows compile error (ambiguous call to math function 'sqrt') and a related compile warning.

+++ libEM/aligner.cpp
@@ -2698,7 +2698,7 @@ vector<Dict> RT3DTreeAligner::xform_align_nbest(EMData * this_img, EMData * to, 
 		
 		// This is drawn from single particle analysis testing, which in that case insures that enough sampling to
 		// reasonably fill Fourier space is achieved
-		float astep=89.99/ceil(90.0*9.0/(8*sqrt(4300/ss)));	// 8 is (3+speed) from SPA with speed=5
+		float astep = (float)(89.99/ceil(90.0*9.0/(8.0*sqrt((float)(4300.0/ss)))));	// 8 is (3+speed) from SPA with speed=5
 		
 		// This is for the first loop, we do a full search in a heavily downsampled space
 		if (s_coverage[0]==0.0f) {
